+++ docs/changelog.rst
@@ -19,6 +19,9 @@ New fields for existing flavors:
 - Added `EircodeField` in IE flavor
   (`gh-366 <https://github.com/django/django-localflavor/pull/366>`_).
 
+- Added Models for Spain (`ESPostalCodeField` and `ESIdentityCardNumberField`)
+  (`gh-357 <https://github.com/django/django-localflavor/pull/357>`_)
+
 Modifications to existing flavors:
 
 - Deprecated `generic.checksums.luhn` and `generic.checksums.ean`. Please use the python-stdnum library instead.